Revert "Multigrid base class"
