+++ README.md
@@ -2,15 +2,13 @@
 
 A simple 1D multigrid solver in Python for testing out ideas quickly.
 
-It should be fairly easy to adjust various aspects of the code.  Here are some
-suggestions of things that could be easily looked at with the current version
-of the code:
+It should be fairly easy to adjust various aspects of the code.  Some
+potentially interesting things to look at include:
 * Adjusting the number of levels or number of multigrid V/W cycles.
 * Turning the coloring of the smoother on/off or use more colors.
 * Adjusting the number of smoothing steps per multigrid cycle.
 * Changing between a V- and W- cycle.
-* Making the interpolation operators problem-dependent.  (This will require
-some effort to code up your own interpolation matrix.)
+* Making the interpolation operators problem-dependent.
 
 Some potential to-do's:
 * Implement a 2D/3D version.
