+++ example/example_1D.py
@@ -1,11 +1,8 @@
-""" Sample 1D homogeneous problem with zero boundary conditions. """
-
 import sys
-sys.path.insert(0, '..')
+sys.path.insert(0,'..')
 
 from py_simplemg import MultigridOptions, SmootherOptions, solve_multigrid
 import numpy as np
-
 nx = 2**8+1
 A = np.zeros((nx, nx))
 b = np.zeros(nx)
@@ -19,6 +16,6 @@ for i in range(nx):
 x = np.zeros(nx)
 x[0:nx//2] = 0.5
 
-my_mg_opts = MultigridOptions(num_its=1, num_levels=4, cycle='W', geom_type='1D')
+my_mg_opts = MultigridOptions(num_its=1, num_levels=4, cycle='W')
 my_smooth_opts = SmootherOptions(smoothdown=1, smoothup=0, redblack=True)
 print(solve_multigrid(A, b, x, my_mg_opts, my_smooth_opts))
