+++ py_simplemg/__init__.py
@@ -1,6 +1,5 @@
 
 from __future__ import absolute_import
-from .multigrid_base import MultigridOptions
-from .multigrid_solver import solve_multigrid
-from .smoothers import SmootherOptions
+from .multigrid_1D import MultigridOptions, SmootherOptions, MultigridLevel, solve_multigrid
+
